refactor: profile page decomposition
